fix: allow multiple mig-vm runners in the same project
This change will allow multiple runners to be registered in the same project, by setting the `instance_name` input variable to different values for each instance.

When applying this change in an existing environment, the secret `gh-token` will be destroyed and recreated with a new name that is based on the `instance_name` variable. Likewise, the default service account assumed by instances in the MIG will change from `runner-service-account` to an `instance_name` variable-based value.